requirements.txt
Django==3.2.6
djangorestframework==3.12.2
psycopg2==2.9.1
pillow==7.2.0
flake8>=3.6.0,<3.7.0